requirements.txt
alpaca-py>=0.21.0
anthropic>=0.40.0
pytest>=8.0.0
pytest-mock>=3.12.0
python-dotenv>=1.0.0
pytz>=2024.1
